Move CPenaltyState traits
