+++ include/football/CPenaltyState.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "football/CChanceState.h"
+#include "football/traits/CPenaltyState.h"
 
 #include "football/CMatchConfiguration.h"
 #include "football/CPlayer.h"
@@ -8,27 +9,7 @@
 
 #include "ExceptionUtils.h"
 
-namespace futsim
-{
-
-namespace football
-{
-class CPenaltyState;
-}
-
-template <> struct json_traits<football::CPenaltyState>
-{
-	//! JSON key for the class.
-	static inline constexpr std::string_view KEY = "Penalty state";
-	//! JSON key for the \copybrief football::CPenaltyState::mOutcome
-	static inline constexpr std::string_view OUTCOME_KEY = json_traits<football::CChanceState>::OUTCOME;
-	//! JSON key for the \copybrief football::CPenaltyState::mShooter
-	static inline constexpr std::string_view SHOOTER_KEY = json_traits<football::CChanceState>::ACTOR<football::E_PLAYER_SKILL::Sh>;
-	//! JSON key for the \copybrief football::CPenaltyState::mStopper
-	static inline constexpr std::string_view STOPPER_KEY = json_traits<football::CChanceState>::ACTOR<football::E_PLAYER_SKILL::St>;
-};
-
-namespace football
+namespace futsim::football
 {
 
 /**
@@ -92,6 +73,4 @@ CPenaltyState::CPenaltyState(
 }
 FUTSIM_CATCH_AND_RETHROW_EXCEPTION( std::invalid_argument, "Error creating the penalty state." )
 
-} // football namespace
-
-} // futsim namespace
+} // futsim::football namespace
